Remove CountVisibleNode redundant code
