--- src/test/java/com/codility/count_visible_nodes/TestCase.java
@@ -1,7 +1,0 @@
-package com.codility.count_visible_nodes;
-
-public interface TestCase 
-{
-  public Tree getCase();
-  public int  getReturn();
-}
