--- src/test/java/com/codility/count_visible_nodes/TestCase01.java
@@ -1,34 +1,0 @@
-package com.codility.count_visible_nodes;
-
-public class TestCase01 implements TestCase 
-{
-  @Override
-  public Tree getCase()
-  {
-    Tree T = new Tree();
-    T.x = 5;
-
-    T.l = new Tree();
-    T.l.x = 3;
-
-    T.r = new Tree();
-    T.r.x = 10;
-    
-    T.l.l = new Tree();
-    T.l.l.x = 20;
-
-    T.l.r = new Tree();
-    T.l.r.x = 21;
-
-    T.r.l = new Tree();
-    T.r.l.x = 1;
-    
-    return T;
-  }
-
-  @Override
-  public int getReturn() {
-    return 4;
-  }
-
-}
